Explicitly test the event_date_time.
With the new XML fixtures, we need to explicitly set the
event_date_time field with a datestring that does have millliseconds,
and with a datetime string that does not. The default fixture will
contain a datetime string with milliseconds, but we do not want to be
implicit in our test setup.