Get build back to passing
A correct version of lxml is key here. The pinned version doesn't work
because it conflicts with codalib dependencies, but the most recent
version doesn't work either. This seems good.

This isn't 100% correct because the pipenv situation means that tox is
no longer actually installing different versions of Django in different
environments - they get overwritten by pipenv. Need to work through
that.