Force pytz to update
Although tox was succeeding on localhost, it was failing on Travis
due to issues with pytz 2018.4. As pytz is installed by a dependency
we don't have direct control over its version number - we have to add
it to the Pipfile to specify a later version. This forces the most
recent pytz to be in the Pipfile.lock.